# spring/classes/openapi/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Abacus WEB Sample",
    "description": "Abacus WEB Sample Project",
    "license": {
      "name": "Apache 2.0",
      "url": "http://abacus.co.kr"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://192.168.2.142:8080/api/",
      "description": "dev"
    }
  ],
...
  "paths": {
...
    "/company/updateCompany": {
      "post": {
        "tags": [
          "서비스 등록 - 고객사/제조사 관리"
        ],
        "summary": "고객사/제조사 수정",
        "description": "고객사/제조사 수정",
        "operationId": "updateCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpCd": {
                    "type": "string",
                    "example": 1
                  },
                  "cmpType": {
                    "type": "string",
                    "example": 10
                  },
                  "cmpNm": {
                    "type": "string",
                    "example": "테스트 회사"
                  },
                  "cmpHomepage": {
                    "type": "string",
                    "example": "http://test.co.kr"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/company/retrieveCompany": {
      "post": {
        "tags": [
          "서비스 등록 - 고객사/제조사 관리"
        ],
        "summary": "고객사/제조사 상세",
        "description": "고객사/제조사 상세",
        "operationId": "retrieveCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpCd": {
                    "type": "string",
                    "example": 1
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponseCompany"
                }
              }
            }
          }
        }
      }
    },
    "/company/retrieveCompanyList": {
      "post": {
        "tags": [
          "서비스 등록 - 고객사/제조사 관리"
        ],
        "summary": "고객사/제조사 목록 조회",
        "description": "고객사/제조사 목록 조회",
        "operationId": "retrieveCompanyList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpType": {
                    "type": "string",
                    "example": 10
                  },
                  "cmpNm": {
                    "type": "string",
                    "example": ""
                  },
                  "currentPage": {
                    "type": "string",
                    "example": 1
                  },
                  "limit": {
                    "type": "string",
                    "example": 10
                  },
                  "pageYn": {
                    "type": "string",
                    "example": "Y"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponseCompanyList"
                }
              }
            }
          }
        }
      }
    },
    "/company/downloadCompanyList": {
      "post": {
        "tags": [
          "서비스 등록 - 고객사/제조사 관리"
        ],
        "summary": "고객사/제조사 엑셀 다운로드",
        "description": "고객사/제조사 엑셀 다운로드",
        "operationId": "downloadCompanyList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpType": {
                    "type": "string",
                    "example": 10
                  },
                  "cmpNm": {
                    "type": "string",
                    "example": ""
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/company/deleteCompany": {
      "post": {
        "tags": [
          "서비스 등록 - 고객사/제조사 관리"
        ],
        "summary": "고객사/제조사 삭제",
        "description": "고객사/제조사 삭제",
        "operationId": "deleteCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpCd": {
                    "type": "string",
                    "example": 1
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/company/createCompany": {
      "post": {
        "tags": [
          "서비스 등록 - 고객사/제조사 관리"
        ],
        "summary": "고객사/제조사 등록",
        "description": "고객사/제조사 등록",
        "operationId": "createCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpType": {
                    "type": "string",
                    "example": 10
                  },
                  "cmpNm": {
                    "type": "string",
                    "example": "테스트 회사"
                  },
                  "cmpHomepage": {
                    "type": "string",
                    "example": "https://test.co.kr"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/common/page-access-check": {
      "post": {
        "tags": [
          "페이지 접근 권한 제어"
        ],
        "operationId": "checkPageAccess",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComResponseCompany": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/Company"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
      "Company": {
        "required": [
          "cmpCd",
          "cmpHomepage",
          "cmpNm",
          "cmpType"
        ],
        "type": "object",
        "properties": {
          "useYn": {
            "type": "string"
          },
          "creDt": {
            "type": "string"
          },
          "creUsr": {
            "type": "string"
          },
          "updDt": {
            "type": "string"
          },
          "updUsr": {
            "type": "string"
          },
          "rolNm": {
            "type": "string"
          },
          "cmpCd": {
            "pattern": "^[0-9]+$",
            "type": "string",
            "description": "회사 코드",
            "example": "1"
          },
          "cmpType": {
            "maxLength": 20,
            "minLength": 0,
            "pattern": "^(10|20)?$",
            "type": "string",
            "description": "회사 구분 코드",
            "example": "10"
          },
          "cmpNm": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "description": "회사명",
            "example": "테스트 회사"
          },
          "cmpHomepage": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "홈페이지",
            "example": "https://test.co.kr"
          },
          "svcNm": {
            "type": "string",
            "description": "서비스명",
            "example": "테스트 서비스"
          },
          "rawCmpType": {
            "type": "string",
            "readOnly": true
          }
        }
      },
      "ComResponseCompanyList": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/CompanyList"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
      "CompanyList": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Company"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AccessAuthRequest": {
        "type": "object",
        "properties": {
          "menuUrl": {
            "type": "string"
          }
        }
      },
...
    }
  }
}