# spring/generated-sources/openapi/src/main/java/com/openapi/generator/aicc/api/CompanyApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.openapi.generator.aicc.api;

import com.openapi.generator.aicc.dto.ComResponseCompany;
import com.openapi.generator.aicc.dto.ComResponseCompanyList;
import com.openapi.generator.aicc.dto.CreateCompanyRequest;
import com.openapi.generator.aicc.dto.DownloadCompanyListRequest;
import com.openapi.generator.aicc.dto.RetrieveCompanyListRequest;
import com.openapi.generator.aicc.dto.RetrieveCompanyRequest;
import com.openapi.generator.aicc.dto.UpdateCompanyRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T16:30:08.398412+09:00[Asia/Seoul]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "서비스 등록 - 고객사/제조사 관리", description = "서비스 등록 - 고객사/제조사 관리")
public interface CompanyApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_COMPANY = "/company/createCompany";
    /**
     * POST /company/createCompany : 고객사/제조사 등록
     * 고객사/제조사 등록
     *
     * @param createCompanyRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createCompany",
        summary = "고객사/제조사 등록",
        description = "고객사/제조사 등록",
        tags = { "서비스 등록 - 고객사/제조사 관리" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CompanyApi.PATH_CREATE_COMPANY,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> createCompany(
        @Parameter(name = "CreateCompanyRequest", description = "", required = true) @Valid @RequestBody CreateCompanyRequest createCompanyRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_COMPANY = "/company/deleteCompany";
    /**
     * POST /company/deleteCompany : 고객사/제조사 삭제
     * 고객사/제조사 삭제
     *
     * @param retrieveCompanyRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteCompany",
        summary = "고객사/제조사 삭제",
        description = "고객사/제조사 삭제",
        tags = { "서비스 등록 - 고객사/제조사 관리" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CompanyApi.PATH_DELETE_COMPANY,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> deleteCompany(
        @Parameter(name = "RetrieveCompanyRequest", description = "", required = true) @Valid @RequestBody RetrieveCompanyRequest retrieveCompanyRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DOWNLOAD_COMPANY_LIST = "/company/downloadCompanyList";
    /**
     * POST /company/downloadCompanyList : 고객사/제조사 엑셀 다운로드
     * 고객사/제조사 엑셀 다운로드
     *
     * @param downloadCompanyListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "downloadCompanyList",
        summary = "고객사/제조사 엑셀 다운로드",
        description = "고객사/제조사 엑셀 다운로드",
        tags = { "서비스 등록 - 고객사/제조사 관리" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CompanyApi.PATH_DOWNLOAD_COMPANY_LIST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> downloadCompanyList(
        @Parameter(name = "DownloadCompanyListRequest", description = "", required = true) @Valid @RequestBody DownloadCompanyListRequest downloadCompanyListRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_COMPANY = "/company/retrieveCompany";
    /**
     * POST /company/retrieveCompany : 고객사/제조사 상세
     * 고객사/제조사 상세
     *
     * @param retrieveCompanyRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveCompany",
        summary = "고객사/제조사 상세",
        description = "고객사/제조사 상세",
        tags = { "서비스 등록 - 고객사/제조사 관리" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseCompany.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CompanyApi.PATH_RETRIEVE_COMPANY,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseCompany> retrieveCompany(
        @Parameter(name = "RetrieveCompanyRequest", description = "", required = true) @Valid @RequestBody RetrieveCompanyRequest retrieveCompanyRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"cmpCd\" : \"1\", \"cmpNm\" : \"테스트 회사\", \"rolNm\" : \"rolNm\", \"svcNm\" : \"테스트 서비스\", \"cmpHomepage\" : \"https://test.co.kr\", \"useYn\" : \"useYn\", \"rawCmpType\" : \"rawCmpType\", \"creUsr\" : \"creUsr\", \"updUsr\" : \"updUsr\", \"cmpType\" : \"10\", \"updDt\" : \"updDt\", \"creDt\" : \"creDt\" }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_COMPANY_LIST = "/company/retrieveCompanyList";
    /**
     * POST /company/retrieveCompanyList : 고객사/제조사 목록 조회
     * 고객사/제조사 목록 조회
     *
     * @param retrieveCompanyListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveCompanyList",
        summary = "고객사/제조사 목록 조회",
        description = "고객사/제조사 목록 조회",
        tags = { "서비스 등록 - 고객사/제조사 관리" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseCompanyList.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CompanyApi.PATH_RETRIEVE_COMPANY_LIST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseCompanyList> retrieveCompanyList(
        @Parameter(name = "RetrieveCompanyListRequest", description = "", required = true) @Valid @RequestBody RetrieveCompanyListRequest retrieveCompanyListRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"list\" : [ { \"cmpCd\" : \"1\", \"cmpNm\" : \"테스트 회사\", \"rolNm\" : \"rolNm\", \"svcNm\" : \"테스트 서비스\", \"cmpHomepage\" : \"https://test.co.kr\", \"useYn\" : \"useYn\", \"rawCmpType\" : \"rawCmpType\", \"creUsr\" : \"creUsr\", \"updUsr\" : \"updUsr\", \"cmpType\" : \"10\", \"updDt\" : \"updDt\", \"creDt\" : \"creDt\" }, { \"cmpCd\" : \"1\", \"cmpNm\" : \"테스트 회사\", \"rolNm\" : \"rolNm\", \"svcNm\" : \"테스트 서비스\", \"cmpHomepage\" : \"https://test.co.kr\", \"useYn\" : \"useYn\", \"rawCmpType\" : \"rawCmpType\", \"creUsr\" : \"creUsr\", \"updUsr\" : \"updUsr\", \"cmpType\" : \"10\", \"updDt\" : \"updDt\", \"creDt\" : \"creDt\" } ], \"totalCount\" : 6 }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_COMPANY = "/company/updateCompany";
    /**
     * POST /company/updateCompany : 고객사/제조사 수정
     * 고객사/제조사 수정
     *
     * @param updateCompanyRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateCompany",
        summary = "고객사/제조사 수정",
        description = "고객사/제조사 수정",
        tags = { "서비스 등록 - 고객사/제조사 관리" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CompanyApi.PATH_UPDATE_COMPANY,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> updateCompany(
        @Parameter(name = "UpdateCompanyRequest", description = "", required = true) @Valid @RequestBody UpdateCompanyRequest updateCompanyRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}